pdf_processor/rag_chain.py
from typing import List, Dict, Optional
import logging
from dataclasses import dataclass
import httpx
import os
from django.conf import settings
from .embeddings import VectorStore

logger = logging.getLogger(__name__)

@dataclass
class QueryResult:
    """Data class to hold query result information"""
    response: str
    source_documents: List[Dict]
    plan_ids: List[str]
    confidence_score: float

class RAGChain:
    """
    Handles the RAG (Retrieval Augmented Generation) pipeline for Medicare plan queries.
    """
    
    def __init__(self):
        """Initialize the RAG pipeline with vector store and LLM configuration."""
        self.vector_store = VectorStore()
        self.ollama_base_url = f"http://{os.getenv('OLLAMA_HOST', 'localhost')}:{os.getenv('OLLAMA_PORT', '11434')}"
        
        # Template for structuring prompts
        self.prompt_template = """
        Answer the following question about Medicare plans based on the provided context.
        Only use information from the context. If you cannot find the information in the context,
        say that you cannot find that specific information.

        Context:
        {context}

        Question:
        {question}

        Please format your response in a clear, direct manner. If comparing multiple plans,
        structure the response to clearly show the differences.
        """

    async def query(
        self,
        question: str,
        plan_ids: Optional[List[str]] = None,
        max_sources: int = 5
    ) -> QueryResult:
        """
        Process a query through the RAG pipeline.

        Args:
            question (str): User's question
            plan_ids (Optional[List[str]]): Specific plan IDs to query
            max_sources (int): Maximum number of sources to retrieve

        Returns:
            QueryResult: Query response with sources and metadata
        """
        try:
            # TODO: Implement RAG pipeline:
            # 1. Retrieve relevant documents
            # 2. Format prompt with context
            # 3. Generate response
            # 4. Format and validate response
            pass
        except Exception as e:
            logger.error(f"Error processing RAG query: {str(e)}")
            raise RAGException(f"Failed to process query: {str(e)}")

    async def _retrieve_relevant_documents(
        self,
        question: str,
        plan_ids: Optional[List[str]] = None,
        max_sources: int = 5
    ) -> List[Dict]:
        """
        Retrieve relevant documents from the vector store.

        Args:
            question (str): User's question
            plan_ids (Optional[List[str]]): Specific plan IDs to query
            max_sources (int): Maximum number of sources to retrieve

        Returns:
            List[Dict]: Relevant documents with metadata
        """
        # TODO: Implement document retrieval
        pass

    async def _generate_llm_response(
        self,
        formatted_prompt: str
    ) -> str:
        """
        Generate response using Ollama LLM.

        Args:
            formatted_prompt (str): Formatted prompt with context

        Returns:
            str: LLM response
        """
        # TODO: Implement LLM call
        pass

    def _format_prompt(
        self,
        question: str,
        relevant_docs: List[Dict]
    ) -> str:
        """
        Format the prompt with question and context.

        Args:
            question (str): User's question
            relevant_docs (List[Dict]): Retrieved relevant documents

        Returns:
            str: Formatted prompt
        """
        # TODO: Implement prompt formatting
        pass

    def _calculate_confidence(
        self,
        question: str,
        response: str,
        source_docs: List[Dict]
    ) -> float:
        """
        Calculate confidence score for the response.

        Args:
            question (str): Original question
            response (str): Generated response
            source_docs (List[Dict]): Source documents used

        Returns:
            float: Confidence score between 0 and 1
        """
        # TODO: Implement confidence calculation
        pass

    def _validate_response(
        self,
        response: str,
        source_docs: List[Dict]
    ) -> bool:
        """
        Validate that the response is grounded in the source documents.

        Args:
            response (str): Generated response
            source_docs (List[Dict]): Source documents

        Returns:
            bool: Whether the response is valid
        """
        # TODO: Implement response validation
        pass

class RAGException(Exception):
    """Custom exception for RAG pipeline errors"""
    pass