requirements.txt
django==5.0.2
djangorestframework==3.14.0
PyPDF2==3.0.1
python-magic==0.4.27
chromadb==0.4.22
sentence-transformers==2.3.1
ollama==0.1.6
python-dotenv==1.0.1
pydantic==2.6.1
pytest==8.0.0
pytest-django==4.8.0
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
